Backend/huxgym/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cssutils==2.3.0
Django==3.2.3
django-cors-headers==3.7.0
django-inlinecss==0.3.0
django-naomi==0.8
djangorestframework==3.12.4
drf-base64==2.0
drf-yasg==1.20.0
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pynliner==0.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
drf_yasg==1.20.0